ml/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.20
qdrant-client==1.14.3
python-dotenv==1.1.1
faiss-cpu==1.11.0
sentence-transformers==5.0.0
torch==2.7.1
torchvision==0.22.1
numpy==2.3.1
requests==2.32.4
pillow==11.3.0
pandas==2.3.1
tqdm==4.67.1